move mongoose.js to database folder, modify server.js to correctly import database
